// src_old/interfaces.ts
export interface TableColumn<T> {
  field: keyof T;
  label: string;
  format?: (item: T[keyof T][keyof T[keyof T]] | T[keyof T]) => React.ReactNode;
}

export interface PaginatedResponse<T> {
  data: T[];
  total: number;
  page: number;
  limit: number;
}

export interface GenericTableProps<T> {
  tableName: string;
  columns: TableColumn<T>[];
  fetchData: (
    page: number,
    pageSize: number,
    searchQuery: string
  ) => Promise<PaginatedResponse<T>>;
}

export interface ActionButton {
  icon: React.ReactNode;
  onClick: (id: number, item?: any) => void;
  title?: string;
  disabled?: boolean;
}

export interface SearchBarProps {
  onSearch: (value: string) => void;
}

export interface AdditionalColumn<T> {
  type: 'radio' | 'checkbox';
  onChange: (itemId: number, checked: boolean) => void;
  valueGetter: (item: T) => boolean;
  columnName: string;
}